Add student class and refactor class main and app
