+++ main.rb
@@ -1,45 +1,11 @@
 require_relative 'app'
-require './create_person'
 
 class Main
-  def initialize
-    @app = App.new
-    @new_person = Create_Person.new
-    @output = 0
-  end
-
-  def option
-    print 'Welcome to School Library App!'
-    while @output != 7
-      print "\nPlease choose an option by entering a number: \n\n"
-      print "1 - List all books\n"
-      print "2 - List all people\n"
-      print "3 - Create a person\n"
-      print "4 - Create a book\n"
-      print "5 - Create a rental\n"
-      print "6 - List all rentals for a given person id\n"
-      print "7 - Exit\n"
-      start_app
-    end
-  end
-
-  def start_app
-    @output = gets.chomp.to_i
-    case @output
-    when 1
-      @app.list_books
-    when 2
-      @new_person.list_people
-    when 3
-      @new_person.create_person
-    when 4
-      @app.create_new_book
-    when 5
-      @app.create_rental
-    when 6
-      @app.rent_list_by_id
-    end
+  def run
+    app = App.new
+    app.options
   end
 end
+
 main = Main.new
-main.option
+main.run
